requirements.txt
# Requisitos para Detecção de Pessoas em Esportes
# Baseado no YOLOv8 e OpenCV

# Core dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0

# Data processing
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=8.0.0

# Utilities
PyYAML>=6.0
pathlib2>=2.3.0
tqdm>=4.60.0

# Optional: for better performance
# CUDA support (uncomment if you have CUDA)
# torch-audio>=0.9.0
# torchaudio>=0.9.0

# Development dependencies (optional)
# jupyter>=1.0.0
# ipython>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
